% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2d_function.R
\name{count2d}
\alias{count2d}
\title{Count the number of observation in 2d space.}
\usage{
count2d(
  x,
  y,
  begin = NA,
  end = NA,
  len = NA,
  xbegin = NA,
  xend = NA,
  xlen = NA,
  ybegin = NA,
  yend = NA,
  ylen = NA
)
}
\arguments{
\item{x}{A numeric vector with the x positions.}

\item{y}{A numeric vector with the y positions.}

\item{begin}{A numeric value to begin the x and y bins.}

\item{end}{A numeric value to end the x and y bins.}

\item{len}{A numeric value to use and the x and y bin width.}

\item{xbegin}{A numeric value to begin the x bins.}

\item{xend}{A numeric value to end the x bins.}

\item{xlen}{A numeric value to use and the x bin width.}

\item{ybegin}{A numeric value to begin the y bins.}

\item{yend}{A numeric value to end the y bins.}

\item{ylen}{A numeric value to use and the y bin width.}
}
\value{
A dataframe with the x and y bin number, centers, range, and the count.
}
\description{
This function will count the number of observations
in 2d space with the ability to set the x and y begin,
end, and width(len) values. Provide either begin, end, and len,
or xbegin, xend, xlen, ybegin, yend, and ylen.
}
\examples{

\donttest{
### Create Test data
x1 <- runif(1000,0,10)
y1 <- runif(1000,0,10)
data1 <- data.frame(x=x1,y=y1)

### Count the number of observations from 0 to 10 by 1.
dfout <- count2d(x=data1$x,y=data1$y,begin=0,end=10,len=1)
# or
dfout <- count2d(x=data1$x,y=data1$y,
                 xbegin=0,xend=10,xlen=1,
                 ybegin=0,yend=10,ylen=1)
}


}
