% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq_function.R
\name{ineq}
\alias{ineq}
\title{An Inequality Function}
\usage{
ineq(x, val, eq)
}
\arguments{
\item{x}{Data to pass through function. Vector with multiple values}

\item{val}{Value to evaluate with inequality}

\item{eq}{inequality symbol as a character.
Options are ">", "<", ">=", "<=", or "=".}
}
\value{
Value of observations that fit the condition given.
}
\description{
This function allows the user to count the number of values/rows
that match the inequality.
}
\examples{
### Create Test data
a<-seq(1,5,by=1)
b<-seq(1,10,by=2)
c<-seq(1,15,by=3)
data<-data.frame(a,b,c)

#for vector:
ineq(a,val=3,eq="<")

#for dataframe:
apply(data,2,function(x) ineq(x,val=4,eq=">"))
}
