% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcorr_filter_function.R
\name{rcorr_filter}
\alias{rcorr_filter}
\title{rcorr filter}
\usage{
rcorr_filter(rcorr_obj, signif = NA, a = 0.05, ccdir = "all", min_abs_cc = NA)
}
\arguments{
\item{rcorr_obj}{The rcorr list.}

\item{signif}{A logical value to filter the p-value by based on a alpha of a. Default is set to NA.}

\item{a}{The numeric alpha value used to filter by significance. Default is set to 0.05.}

\item{ccdir}{A character string of the correlation coefficient direction used to filter. Options are "all", "zero", "pos", or "neg".}

\item{min_abs_cc}{A numeric minimum absolute value of the correlation coefficient to return.}
}
\value{
A data.frame indicating the pairs of variables with correlation coefficients or p-values which match the conditions provided.
}
\description{
This function takes an rcorr list, converts in into a data.frame, and allows the user to filter the correlations based on significance and correlation coefficient.
}
\examples{
library(tidyverse)
library(Hmisc)

# create rcorr object
out <- mtcars \%>\%
  select(mpg:qsec) \%>\%
  as.matrix() \%>\%
  rcorr()

# filter by signif
rcorr_filter(rcorr_obj = out, signif = NA, a = 0.05, ccdir = "all", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "all", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = FALSE, a = 0.05, ccdir = "all", min_abs_cc = NA)

# with different alpha values
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "all", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.01, ccdir = "all", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.001, ccdir = "all", min_abs_cc = NA)

# filter by ccdir
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "all", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "pos", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "neg", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "zero", min_abs_cc = NA)

# filter by min_abs_cc
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "all", min_abs_cc = NA)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "all", min_abs_cc = 0.5)
rcorr_filter(rcorr_obj = out, signif = TRUE, a = 0.05, ccdir = "pos", min_abs_cc = 0.5)
}
