% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reextent_function.R
\name{reextent}
\alias{reextent}
\title{A function to alter the extent of rasters}
\usage{
reextent(ras, method)
}
\arguments{
\item{ras}{Raster object to shift the extent}

\item{method}{Numerical value to indicate direction of
extent change. '1' shifts from -180:180 to 0:360. '2' shifts from
0:360 to -180:180.}
}
\value{
Shifted raster
}
\description{
This function shifts the extent of a raster from
 -180:180 to 0:360 and vice versa
}
\examples{
### Create Test data
r <- raster::raster(vals = c(rep(1:360,length.out=64800/2),rep(1:360,length.out=64800/2)))
plot(r)
# convert r from -180:180 to 0:360
out <- reextent(r, method = 1)
plot(out)
# convert back to -180:180
out2 <- reextent(out, method = 2)
plot(out2)

}
