% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_select_function.R
\name{select_select}
\alias{select_select}
\title{Select which columns to put in dplyr::select()}
\usage{
select_select(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\description{
Opens a shiny app to help select and sort desired columns.
Drag and drop levels in the appropriate order.
Copy the code output and replace the select_select() call.
}
\examples{
\donttest{mtcars \%>\% select_select()}

}
