% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_sort_function.R
\name{factor_sort}
\alias{factor_sort}
\title{Drag and drop to reorder levels of a factor.}
\usage{
factor_sort(data, col)
}
\arguments{
\item{data}{A data.frame}

\item{col}{A factor or character vector}
}
\description{
Opens a shiny app to help resort levels of a factor.
Drag and drop levels in the appropriate order.
Copy the code output and replace the factor_sort() call.
}
\examples{
\donttest{chickwts \%>\% factor_sort(col = feed)}


}
